Implemented ability to get accurate costs based on Ingredient.java amount and individual ingredient cost.
Each Brew now has a cloned version of the Ingredient.java it is based on, allowing for use of the amount field (this might be worth cleaning up in future to have some sort of separate class for the IngredientManager.java map, but for now it isn't causing any issues as far as is seen.

Additionally, Brew.getOwner and functionality required to save and load the data has been implemented.