Documentation for the MainWindowController.java completed.
In addition, refactored MainWindowController.setEditPropertiesOnOwnerColumn to focus solely on the Owner Column. A more general version of this method may be needed in the future (Perhaps in some sort of util class).

TODO added to reflect this possibility.

Finally MainWindowController.initialize has had the lambda functionality of the brewFactory removed (legacy).