Documentation for the IngredientMenuWindowController.java completed.
In addition, refactored IngredientMenuWindowController.setEditPropertiesOnColumn to focus solely on the Ingredient Cost Column. A more general version of this method may be needed in the future (Perhaps in some sort of util class).

TODO added to reflect this possibility.
TODO also added to remind that the addButton requires functionality within this class (The purpose of which is to allow users to add new ingredients to the list)